Replace custom client with Grafana SDK
This means a bit less boilerplate in here. The SDK isn't the most mature
so keeping the httpmock tests is probably wise - the main difference is
that the SDK puts the auth in the URL rather than the headers so the
basic auth wrapper was no longer needed. Did some refactoring to share
some of the code around creating the annotation text between outCreate
and outUpdate

Signed-off-by: Andy Paine <andy.paine@engineerbetter.com>